only allow authorized users to delete threads

+++ app/Http/Controllers/Threads/ThreadsController.php
@@ -136,6 +136,13 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+
+        $this->authorize('update', $thread);
+
+        /*if($thread->user_id != auth()->id()) {
+            abort(403, 'You do not have permission to do this.');
+        }*/
+
         $thread->delete();
 
         if(\request()->wantsJson())
